# apps/api/openapi/types.gen.go
// Code generated by packages/contracts/codegen.ts; DO NOT EDIT.
package openapi

import "time"

type Book struct {
    ID        string    `json:"id"`
    Title     string    `json:"title"`
    Author    string    `json:"author"`
    Price     float64   `json:"price"`
    Currency  string    `json:"currency"`
    Stock     int       `json:"stock"`
    CreatedAt time.Time `json:"created_at"`
    UpdatedAt time.Time `json:"updated_at"`
}

type BookResponse struct {
    Data Book `json:"data"`
}

type BookListResponse struct {
    Data       []Book  `json:"data"`
    NextCursor *string `json:"next_cursor,omitempty"`
}

type BookCreateRequest struct {
    Title    string  `json:"title"`
    Author   string  `json:"author"`
    Price    float64 `json:"price"`
    Currency string  `json:"currency"`
    Stock    int     `json:"stock"`
}

type BookUpdateRequest = BookCreateRequest

type PurchaseRequest struct {
    Quantity   int    `json:"quantity"`
    CustomerID string `json:"customer_id,omitempty"`
}

type PurchaseAcceptedResponse struct {
    MessageID string `json:"message_id"`
    Status    string `json:"status"`
}

type ErrorResponse struct {
    Error   string      `json:"error"`
    Code    string      `json:"code"`
    Details interface{} `json:"details,omitempty"`
}

type HealthResponse struct {
    Status string `json:"status"`
}